requirements.txt
Django~=4.0.3
python-dotenv~=0.20.0
djangorestframework~=3.13.1